# docs/swagger.yaml
basePath: /api
definitions:
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  service.APIKeyCreateRequest:
    properties:
      daily_limit:
        type: integer
      daily_tokens:
        type: integer
      expires_at:
        type: string
      monthly_limit:
        type: integer
      monthly_tokens:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  service.APIKeyUpdateRequest:
    properties:
      active:
        type: boolean
      daily_limit:
        type: integer
      daily_tokens:
        type: integer
      expires_at:
        type: string
      monthly_limit:
        type: integer
      monthly_tokens:
        type: integer
      name:
        type: string
    type: object
  service.CredentialCreateRequest:
    properties:
      api_key:
        type: string
      api_key_id:
        type: integer
      base_url:
        type: string
      default_model:
        type: string
      name:
        type: string
      provider_type:
        type: string
    required:
    - api_key
    - api_key_id
    - base_url
    - name
    - provider_type
    type: object
  service.CredentialUpdateRequest:
    properties:
      active:
        type: boolean
      api_key:
        type: string
      base_url:
        type: string
      default_model:
        type: string
      name:
        type: string
      provider_type:
        type: string
    type: object
...
host: localhost:7847
info:
  contact: {}
  description: LLM 文档处理与知识图谱 API 中间件
  title: Nexus API
  version: "1.0"
paths:
  /admin/api-keys:
    get:
      description: 获取所有 API Key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 API Key 列表
      tags:
      - API Key
    post:
      consumes:
      - application/json
      description: 创建新的 API Key，用于外部接口认证
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.APIKeyCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建 API Key
      tags:
      - API Key
  /admin/api-keys/{id}:
    delete:
      description: 删除指定 API Key
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除 API Key
      tags:
      - API Key
    put:
      consumes:
      - application/json
      description: 更新 API Key 的名称、状态等信息
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.APIKeyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新 API Key
      tags:
      - API Key
  /admin/api-keys/{id}/usage:
    get:
      description: 查询指定 API Key 的使用统计
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - default: 30
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 查询 API Key 用量
      tags:
      - API Key
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名密码登录，返回 JWT Token
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 管理员登录
      tags:
      - 认证
  /admin/credentials:
    get:
      description: 获取凭证列表，可按 API Key 过滤
      parameters:
      - description: API Key ID
        in: query
        name: api_key_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取凭证列表
      tags:
      - 凭证
    post:
      consumes:
      - application/json
      description: 创建 LLM 凭证，绑定到 API Key，用于外部用户自带模型
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CredentialCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建凭证
      tags:
      - 凭证
  /admin/credentials/{id}:
    delete:
      description: 删除指定凭证
      parameters:
      - description: 凭证 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除凭证
      tags:
      - 凭证
    put:
      consumes:
      - application/json
      description: 更新凭证信息
      parameters:
      - description: 凭证 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CredentialUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新凭证
      tags:
      - 凭证
  /admin/entities:
    get:
      description: 分页获取实体，支持按类型和关键词筛选
      parameters:
      - description: 实体类型
        in: query
        name: type
        type: string
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取实体列表
      tags:
      - 实体
  /admin/entities/{id}:
    get:
      description: 获取指定实体
      parameters:
      - description: 实体 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取实体详情
      tags:
      - 实体
  /admin/entities/{id}/relations:
    get:
      description: 获取指定实体的所有关系
      parameters:
      - description: 实体 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取实体关系
      tags:
      - 实体
  /admin/graph:
    get:
      description: 获取实体关系图谱数据
      parameters:
      - default: 200
        description: 节点数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取知识图谱
      tags:
      - 图谱
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"